feat(fe): show loading skeleton on component basis
