+++ frontend-client/app/problem/[problemId]/submission/[submissionId]/page.tsx
@@ -1,37 +1,19 @@
-import Codeeditor from '@/components/Codeeditor'
-import { ScrollArea, ScrollBar } from '@/components/ui/scroll-area'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow
-} from '@/components/ui/table'
-import { fetcherWithAuth } from '@/lib/utils'
-import type { SubmissionDetail } from '@/types/type'
-import dayjs from 'dayjs'
+import { Skeleton } from '@/components/ui/skeleton'
 import { ArrowLeft } from 'lucide-react'
 import Link from 'next/link'
-import { IoIosLock } from 'react-icons/io'
+import { Suspense } from 'react'
+import SubmissionDetail from '../_components/SubmissionDetail'
 
-export default async function SubmissionDetail({
+export default async function Page({
   params
 }: {
   params: {
-    problemId: string
-    submissionId: string
+    problemId: number
+    submissionId: number
   }
 }) {
   const { submissionId, problemId } = params
-  const res = await fetcherWithAuth(`submission/${submissionId}`, {
-    searchParams: {
-      problemId
-    },
-    cache: 'no-store'
-  })
 
-  const submission: SubmissionDetail = await res.json()
   return (
     <div className="flex flex-col gap-5 overflow-auto p-6">
       <div className="flex items-center gap-3 ">
@@ -40,91 +22,17 @@ export default async function SubmissionDetail({
         </Link>
         <h1 className="text-xl font-bold">Submission #{submissionId}</h1>
       </div>
-      {res.ok ? (
-        <>
-          <ScrollArea className="shrink-0 rounded-md">
-            <div className="flex items-center justify-around gap-5 bg-slate-700 p-5 text-sm [&>div]:flex [&>div]:flex-col [&>div]:items-center [&>div]:gap-1 [&_*]:whitespace-nowrap [&_p]:text-slate-400">
-              <div>
-                <h2>User</h2>
-                <p>{submission.username}</p>
-              </div>
-              <div>
-                <h2>Result</h2>
-                <p
-                  className={
-                    submission.result === 'Accepted'
-                      ? '!text-green-500'
-                      : '!text-red-500'
-                  }
-                >
-                  {submission.result}
-                </p>
-              </div>
-              <div>
-                <h2>Language</h2>
-                <p>{submission.language}</p>
-              </div>
-              <div>
-                <h2>Submission Time</h2>
-                <p>
-                  {dayjs(submission.createTime).format('YYYY-MM-DD HH:mm:ss')}
-                </p>
-              </div>
-            </div>
-            <ScrollBar orientation="horizontal" />
-          </ScrollArea>
-          <div>
-            <h2 className="mb-3 text-lg font-bold">Source Code</h2>
-            <Codeeditor
-              value={submission.code}
-              language={submission.language}
-              readOnly
-              className="max-h-96 min-h-16 w-full"
-            />
+      <Suspense
+        fallback={
+          <div className="flex h-fit flex-col gap-4 p-2 text-lg">
+            <Skeleton className="h-20 w-full rounded-lg bg-slate-900" />
+            <Skeleton className="h-8 w-3/12 rounded-lg bg-slate-900" />
+            <Skeleton className="h-32 w-full rounded-lg bg-slate-900" />
           </div>
-          {submission.testcaseResult.length !== 0 && (
-            <div>
-              <h2 className="text-lg font-bold">Test case</h2>
-              <Table className="[&_*]:text-center [&_*]:text-sm [&_*]:hover:bg-transparent [&_td]:p-2 [&_tr]:border-slate-600">
-                <TableHeader className="[&_*]:text-slate-100">
-                  <TableRow>
-                    <TableHead>#</TableHead>
-                    <TableHead>Result</TableHead>
-                    <TableHead>Runtime</TableHead>
-                    <TableHead>Memory</TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {submission.testcaseResult.map((item) => (
-                    <TableRow key={item.id}>
-                      <TableCell>{item.id}</TableCell>
-                      <TableCell
-                        className={
-                          item.result === 'Accepted'
-                            ? 'text-green-500'
-                            : 'text-red-500'
-                        }
-                      >
-                        {item.result}
-                      </TableCell>
-                      <TableCell>{item.cpuTime} ms</TableCell>
-                      <TableCell>{item.memoryUsage} mb</TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </div>
-          )}
-        </>
-      ) : (
-        <div className="flex h-[300px] flex-col items-center justify-center gap-20">
-          <IoIosLock size={100} />
-          <p>
-            Unable to check others&apos; until your correct submission is
-            accepted
-          </p>
-        </div>
-      )}
+        }
+      >
+        <SubmissionDetail problemId={problemId} submissionId={submissionId} />
+      </Suspense>
     </div>
   )
 }
